backend/requirements.txt
pynput
numpy
pandas
scipy
xgboost>=2.0
scikit-learn>=1.3
psutil
joblib
fastapi>=0.104
uvicorn[standard]>=0.24
pydantic>=2.0
shap
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.6
requests>=2.31

# Supabase for extended features
supabase>=2.0.0
python-dotenv>=1.0.0

# LLM integration (optional - graceful degradation if not configured)
openai>=1.0.0

# Composio MCP integration (optional)
composio>=0.1.0
